Make Rubocop happy
Resolved current RuboCop issues.
